代理后台-洗码设置 功能

+++ app/Http/SingleActions/Backend/Merchant/User/UserGrade/DeleteAction.php
@@ -3,7 +3,7 @@
 namespace App\Http\SingleActions\Backend\Merchant\User\UserGrade;
 
 use App\Http\Controllers\BackendApi\BackEndApiMainController;
-use App\Models\User\FrontendUser;
+use App\Models\User\UsersCommissionConfigDetail;
 use App\Models\User\UsersGrade;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\DB;
@@ -33,10 +33,8 @@ class DeleteAction
      * @throws \Exception Exception.
      * @return JsonResponse
      */
-    public function execute(
-        BackEndApiMainController $contll,
-        array $inputDatas
-    ): JsonResponse {
+    public function execute(BackEndApiMainController $contll, array $inputDatas): JsonResponse
+    {
         //检查当前平台的这条数据是否存在。
         $currentUsersGrade = $this->model->where(
             [
@@ -50,18 +48,20 @@ class DeleteAction
         DB::beginTransaction();
         //修改该等级的用户等级ID = null
         if ($currentUsersGrade->user->isNotEmpty()) {
-            $updateUser = FrontendUser::where(
-                [
-                    ['platform_sign', $currentUsersGrade->platform_sign],
-                    ['grade_id', $currentUsersGrade->id],
-                ],
-            )->update(['grade_id' => null]);
+            $users      = $currentUsersGrade->user();
+            $updateUser = $users->update(['grade_id' => null]);
             if (!$updateUser) {
                 DB::rollback();
                 throw new \Exception('200706');
             }
         }
-        // 删除这条等级数据。
+        //删除该等级的洗码设置数据
+        $deleteCommissionDetail = UsersCommissionConfigDetail::where('grade_id', $currentUsersGrade->id)->delete();
+        if (!$deleteCommissionDetail) {
+            DB::rollback();
+            throw new \Exception('200709');
+        }
+        //删除这条等级数据。
         if (!$currentUsersGrade->delete()) {
             DB::rollback();
             throw new \Exception('200705');
